Move enum back to glide3x library version file
