+++ SlashGaming-Diablo-II-Free-Resolution/src/sgd2mapi_extension/glide3x_library_version.hpp
@@ -43,13 +43,21 @@
  *  work.
  */
 
-#ifndef SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_VERSION_HPP_
-#define SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_VERSION_HPP_
+#ifndef SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_VERSION_HPP_
+#define SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_VERSION_HPP_
 
-#include "glide3x_library_version/glide3x_library_version_enum.hpp"
+#include "glide3x_library_version.h"
 
 namespace d2 {
 namespace glide3x_library_version {
+
+enum Glide3xLibraryVersion {
+  kSven1_4_4_21 = ::D2_Glide3xLibraryVersion_kSven1_4_4_21,
+  kSven1_4_6_1 = ::D2_Glide3xLibraryVersion_kSven1_4_6_1,
+  kSven1_4_8_3 = ::D2_Glide3xLibraryVersion_kSven1_4_8_3,
+  kNGlide3_10_0_658 = ::D2_Glide3xLibraryVersion_kNGlide3_10_0_658,
+  kD2dx = ::D2_Glide3xLibraryVersion_kD2dx,
+};
 
 /**
  * Returns the UTF-8 encoded null-terminated string associated with
@@ -71,6 +79,9 @@ Glide3xLibraryVersion GetRunning();
 const char* GetRunningName();
 
 } // namespace glide3x_library_version
+
+typedef glide3x_library_version::Glide3xLibraryVersion Glide3xLibraryVersion;
+
 } // namespace d2
 
-#endif /* SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_VERSION_HPP_ */
+#endif /* SGD2FR_SGD2MAPI_EXTENSION_GLIDE3X_LIBRARY_VERSION_HPP_ */
